pyproject.toml
[tool.poetry]
name = "anderson-liang-sabia-2024-replication"
version = "0.1.0"
description = ""
authors = ["Matheus Alencastro <matheus.alencastro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.3.3"
nbformat = "^5.10.4"
statsmodels = "^0.14.6"
matplotlib = "^3.10.7"
fpdf2 = "^2.8.5"


[tool.poetry.group.dev.dependencies]
ipykernel = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
